backend/requirements.txt
# ============================================================
# ATS Ranker Backend — Python 3.10.9
# ============================================================

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
starlette==0.36.3

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Resume Parsing
pdfplumber==0.10.3
PyMuPDF==1.23.26
python-docx==1.1.0

# NLP & ML
spacy==3.7.4
transformers==4.38.2
sentence-transformers==2.5.1
tokenizers==0.15.2

# Numeric / ML
scikit-learn==1.4.1.post1
numpy==1.26.4
scipy==1.12.0

# PyTorch (CPU)
torch==2.2.1

# HTTP
httpx==0.27.0
requests==2.31.0

# Database
sqlalchemy==2.0.28
aiosqlite==0.20.0
alembic==1.13.1

# File utilities
aiofiles==23.2.1
pillow==10.2.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
